Modularised UI code into different components
