+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "lucide-react": "^0.414.0"
+  }
+}
